test(py): CodSpeed benchmarks, with startup-failure probes
Runs the existing pyperf benchmark cases under pytest-codspeed by sharing one
case list (bench/cases.py) between the pyperf runner and a pytest suite.

The workflow is a faithful clone of the structure from the CodSpeed wizard PR
(#3418), which produced a successful run, since our own hand-written workflow
was rejected by GitHub with startup_failure. The three codspeed-probe-*.yml
files each add exactly one element from our version (paths filters, job-level
defaults.working-directory, mode: simulation,memory) to identify which one is
rejected. They are temporary and get deleted once the culprit is known.
