+++ python/pyproject.toml
@@ -178,6 +178,13 @@ dev = [
     "pytest-timeout>=2.3.1",
 ]
 
+# Benchmarking dependencies (`make benchmark-codspeed`). Kept out of `dev` so
+# it's only installed by the bench workflow / when explicitly requested.
+bench = [
+    "pytest-benchmark>=5.1.0",
+    "pytest-codspeed>=4.3.0",  # 4.3.0+ adds the memory instrument
+]
+
 # Linting dependencies
 lint = [
     "openai>=2.6.0",
